requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.32.0
google-genai>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
kaggle>=1.6.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
